Add error mod
